Move the arrow util module one level up
