+++ python/lsst/dax/ppdb/_arrow.py
