+++ python/lsst/dax/ppdb/export/_chunk_exporter.py
@@ -33,9 +33,9 @@ from lsst.dax.apdb.versionTuple import VersionTuple
 from lsst.ppdb.gcp.auth import get_auth_default
 from lsst.ppdb.gcp.pubsub import Publisher
 
+from .._arrow import write_parquet
 from ..config import PpdbConfig
 from ..sql._ppdb_replica_chunk_sql import ChunkStatus, PpdbReplicaChunkSql
-from ._arrow import write_parquet
 
 __all__ = ["ChunkExporter"]
